Remove intermediate data structure
